pyproject.toml
[project]
name = "gemini-video-trans"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
